% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{KPIs forecasts}
\usage{
forecast(kpi, website, export = FALSE)
}
\arguments{
\item{kpi}{A Google Analytics metric, such as \code{ga:metrics}.}

\item{website}{The unique Google Analytics table ID of the form 
\code{ga:XXXXXXX}, where \code{XXXXXXX} is the Analytics view (profile) ID 
for which the query will retrieve the data.}

\item{export}{If the export option is set as "TRUE", both raw data & graphics 
will be exported in the current working directory.}
}
\description{
The \code{forecast} function makes forecasts for your website KPIs
 for the next 3 months, using time-series.
}
\note{
The black part is the past website traffic data; the blue part is the
graphical representation of the forecast.
}
\examples{
\dontrun{
forecast(kpi = "ga:sessions", website = "ga:XXXXXXX", export = FALSE)}
}

