% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Automatic reports}
\usage{
report(website)
}
\arguments{
\item{website}{The unique Google Analytics table ID of the form 
\code{ga:XXXXXXX}, where \code{XXXXXXX} is the Analytics view (profile) ID 
for which the query will retrieve the data.}
}
\description{
The \code{report} function returns a report presenting your
website KPIs during the last month, and their evolution over a one year period.
}
\note{
The output file will be generated in your working directory. Don't know
 where is your working directory ? Type `getwd()` in the R console.
}
\examples{
\dontrun{
report(website = "ga:XXXXXXX")}
}

