% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekperf.R
\name{weekperf}
\alias{weekperf}
\title{Weekly KPIs assessment}
\usage{
weekperf(kpi, website, export = FALSE)
}
\arguments{
\item{kpi}{A Google Analytics metric, such as \code{ga:metrics}.}

\item{website}{The unique Google Analytics table ID of the form 
\code{ga:XXXXXXX}, where \code{XXXXXXX} is the Analytics view (profile) ID 
for which the query will retrieve the data.}

\item{export}{If the export option is set as "TRUE", both raw data & graphics 
will be exported in the current working directory.}
}
\description{
The \code{weekperf} function returns a plot designed to assess
your website KPIs during the past 7 days.
}
\note{
The triangles are representing the results of the past 7 days.
  Their color may vary according to the mean (green is > mean, red is < mean).
  The mean is represented by the letter \code{m}.
}
\examples{
\dontrun{
weekperf(kpi = "ga:sessions", website = "ga:XXXXXXX", export = TRUE)}
}

